Move TA config parsing to configmap manifest builder
This is in preparation for the Target Allocator CRD. For that CRD, the
Collector instance may be specified via a label or ownership relation.
Therefore, we need to pull the necessary data from the Collector
instance during TA config generation. It's more convenient to do this in
all cases, which is what this change does.
