+++ internal/manifests/collector/targetallocator.go
@@ -15,13 +15,11 @@
 package collector
 
 import (
-	"github.com/mitchellh/mapstructure"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
 	"github.com/open-telemetry/opentelemetry-operator/internal/manifests"
-	"github.com/open-telemetry/opentelemetry-operator/internal/manifests/targetallocator/adapters"
 )
 
 // TargetAllocator builds the TargetAllocator CR for the given instance.
@@ -30,19 +28,6 @@ func TargetAllocator(params manifests.Params) (*v1alpha1.TargetAllocator, error)
 	taSpec := params.OtelCol.Spec.TargetAllocator
 	if !taSpec.Enabled {
 		return nil, nil
-	}
-
-	configStr, err := params.OtelCol.Spec.Config.Yaml()
-	if err != nil {
-		return nil, err
-	}
-	scrapeConfigs, err := getScrapeConfigs(configStr)
-	if err != nil {
-		return nil, err
-	}
-	globalConfig, err := getGlobalConfig(params.OtelCol.Spec.Config)
-	if err != nil {
-		return nil, err
 	}
 
 	return &v1alpha1.TargetAllocator{
@@ -70,50 +55,8 @@ func TargetAllocator(params manifests.Params) (*v1alpha1.TargetAllocator, error)
 			},
 			AllocationStrategy: taSpec.AllocationStrategy,
 			FilterStrategy:     taSpec.FilterStrategy,
-			ScrapeConfigs:      scrapeConfigs,
-			GlobalConfig:       globalConfig,
 			PrometheusCR:       taSpec.PrometheusCR,
 			Observability:      taSpec.Observability,
 		},
 	}, nil
-}
-
-func getGlobalConfig(otelConfig v1beta1.Config) (v1beta1.AnyConfig, error) {
-	// TODO: Eventually we should figure out a way to pull this in to the main specification for the TA
-	type promReceiverConfig struct {
-		Prometheus struct {
-			Config struct {
-				Global map[string]interface{} `mapstructure:"global"`
-			} `mapstructure:"config"`
-		} `mapstructure:"prometheus"`
-	}
-	decodedConfig := &promReceiverConfig{}
-	if err := mapstructure.Decode(otelConfig.Receivers.Object, decodedConfig); err != nil {
-		return v1beta1.AnyConfig{}, err
-	}
-	return v1beta1.AnyConfig{
-		Object: decodedConfig.Prometheus.Config.Global,
-	}, nil
-}
-
-func getScrapeConfigs(otelcolConfig string) ([]v1beta1.AnyConfig, error) {
-	// Collector supports environment variable substitution, but the TA does not.
-	// TA Scrape Configs should have a single "$", as it does not support env var substitution
-	prometheusReceiverConfig, err := adapters.UnescapeDollarSignsInPromConfig(otelcolConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	scrapeConfigs, err := adapters.GetScrapeConfigsFromPromConfig(prometheusReceiverConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	v1beta1scrapeConfigs := make([]v1beta1.AnyConfig, len(scrapeConfigs))
-
-	for i, config := range scrapeConfigs {
-		v1beta1scrapeConfigs[i] = v1beta1.AnyConfig{Object: config}
-	}
-
-	return v1beta1scrapeConfigs, nil
 }
